- add ActTextSetColor - add FocusAdd action - add drawable.setAction - add hasEvent to game - add support to prefixes/suffixes to itemClasses - add zoom label to UI - add zoom support to level - change menu parsing code to support listing items - delete unused utils classes - disable hover on level when hovering over inventory/etc - fix fade effect not resizing - implement descriptions in items - merge setBitmapFont and SetFont - move selectedItem from level to player - refactor ActItem.h - refactor ItemClass parsing to support cloning items - refactor UIObject implementations to use new setAction + parsing code - refactor getProperty into Queryable interface - refactor item code - refactor parseFile to support replaceVars functionality - refactor view (add zoom) - remove item.pickFromLevel (move to loadFromLevel) - update UIObject to add setAction - update item + player related code - update item description code to use namers - update menu parser code (exclude) - update stash to show non equipable items - update switch condition to allow empty actions - update toolset (vs2017) - update variant - update gamefiles (items, zoom on level using scroll, list unidentified items in Cain's panel)

+++ src/SFMLUtils.cpp
@@ -46,7 +46,7 @@ namespace SFMLUtils
 		}
 		else
 		{
-			texRect.left = valX;
+			texRect.left = (int)valX;
 			updateTexRect = true;
 		}
 		float valY = std::round(-((float)texRect.height / 2.f) + ((float)texSize.y / 2.f));
@@ -56,7 +56,7 @@ namespace SFMLUtils
 		}
 		else
 		{
-			texRect.top = valY;
+			texRect.top = (int)valY;
 			updateTexRect = true;
 		}
 		sprite.setOrigin(origin);
